client-api: Merge Error into ruma_common::api::Error
All Matrix endpoint errors should use the same format in theory, and
this type supports other response error formats anyway.
