+++ crates/ruma-client-api/Cargo.toml
@@ -41,7 +41,6 @@ unstable-msc2448 = []
 unstable-msc2654 = []
 unstable-msc3488 = []
 unstable-msc3814 = []
-unstable-msc3843 = []
 unstable-msc3983 = []
 unstable-msc4108 = ["ruma-common/unstable-msc4108"]
 unstable-msc4121 = []
