+++ crates/ruma-client-api/tests/it/uiaa.rs
@@ -2,12 +2,11 @@
 
 use assert_matches2::assert_matches;
 use assign::assign;
-use ruma_client_api::{
-    error::ErrorKind,
-    uiaa::{self, AuthData, AuthFlow, AuthType, UiaaInfo, UiaaResponse, UserIdentifier},
+use ruma_client_api::uiaa::{
+    self, AuthData, AuthFlow, AuthType, UiaaInfo, UiaaResponse, UserIdentifier,
 };
 use ruma_common::{
-    api::{EndpointError, OutgoingResponse},
+    api::{EndpointError, OutgoingResponse, error::ErrorKind},
     canonical_json::assert_to_canonical_json_eq,
 };
 use serde_json::{
