+++ crates/ruma-common/CHANGELOG.md
@@ -41,6 +41,9 @@ Breaking changes:
 - `MatrixError` and `MatrixErrorBody` were renamed to `Error` and `ErrorBody`
   respectively. The `Matrix` prefix is redundant in a crate about the Matrix
   protocol.
+- The `error` module from `ruma-client-api` was merged into the `api::error`
+  module. `Error` is now non-exhaustive and `ErrorBody` has a new `Standard`
+  variant.
 
 Improvements:
