+++ crates/ruma-common/Cargo.toml
@@ -34,6 +34,7 @@ unstable-msc2666 = []
 unstable-msc2870 = []
 unstable-msc3417 = []
 unstable-msc3768 = []
+unstable-msc3843 = []
 unstable-msc3930 = []
 unstable-msc3931 = []
 unstable-msc3932 = ["unstable-msc3931"]
