+++ crates/ruma-common/src/api/error/tests.rs
