+++ crates/ruma-federation-api/Cargo.toml
@@ -23,7 +23,7 @@ client = ["dep:httparse", "dep:memchr"]
 server = ["dep:bytes", "dep:rand"]
 unstable-msc3618 = []
 unstable-msc3723 = []
-unstable-msc3843 = []
+unstable-msc3843 = ["ruma-common/unstable-msc3843"]
 unstable-msc4125 = []
 
 [dependencies]
