+++ crates/ruma/Cargo.toml
@@ -161,7 +161,7 @@ unstable-msc3618 = ["ruma-federation-api?/unstable-msc3618"]
 unstable-msc3723 = ["ruma-federation-api?/unstable-msc3723"]
 unstable-msc3768 = ["ruma-common/unstable-msc3768"]
 unstable-msc3814 = ["ruma-client-api?/unstable-msc3814"]
-unstable-msc3843 = ["ruma-client-api?/unstable-msc3843", "ruma-federation-api?/unstable-msc3843"]
+unstable-msc3843 = ["ruma-common/unstable-msc3843", "ruma-federation-api?/unstable-msc3843"]
 unstable-msc3927 = ["ruma-events?/unstable-msc3927"]
 unstable-msc3930 = ["ruma-common/unstable-msc3930"]
 unstable-msc3931 = ["ruma-common/unstable-msc3931"]
